+++ src/packages/array_basic.rs
@@ -9,6 +9,8 @@ use std::prelude::v1::*;
 use std::{any::TypeId, cmp::Ordering, mem};
 
 def_package!(crate:BasicArrayPackage:"Basic array utilities.", lib, {
+    lib.standard = true;
+
     combine_with_exported_module!(lib, "array", array_functions);
 
     // Register array iterator